requirements.txt
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
